iio: adc: ad4000: Allow config write to non-reg_access devices
We need the write_raw() function to be able to update the sampling
frequency when the platform supports SPI offloading (e.g. SPI-Engine).
That is true even if the ADC does not have a configuration register to
write to. We can also have the ADC wired in a way that the configuration
register is not accesible, but the platform supporting SPI offload.
This is might not be ideal because writes to _scale will be allowed but
never succeed. How to fix this?

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>