iio: adc: ad7768-1: Fix device Synchronization
The synchronization method using GPIO requires the generated pulse to be
truly synchronous with the base MCLK signal. Since this cannot be achieved
in hardware, the datasheet recommends using synchronization over SPI,
where the generated pulse is already synchronous with MCLK. This requires
the SYNC_OUT pin to be connected to SYNC_IN.

To avoid conflicts with the SYNC_OUT, sync-in GPIO is configured as input.

Fixes: e9517dffd1d5 ("iio: adc: ad7768-1: Add support for setting the sampling frequency")
Signed-off-by: Jonathan Santos <Jonathan.Santos@analog.com>